fix(cap): split VoiceTone/VOICE_TONE_LABELS to non-server-only file
lib/cap/voice-tone-labels.ts holds the VoiceTone type and VOICE_TONE_LABELS
constant without server-only so client components can import them directly.
CapSubscriptionPanel ('use client') now imports from voice-tone-labels.
Also includes requireCapOperatorForApi auth-gate registration in audit.ts.
